Rename project from Mjolnir to Hammerspoon.

+++ extensions/build_extensions.sh
@@ -1,18 +1,23 @@
 #!/bin/sh
 
 #  build_extensions.sh
-#  Mjolnir
+#  Hammerspoon
 #
 #  Created by Peter van Dijk on 12/10/14.
 #  Copyright (c) 2014 Steven Degutis. All rights reserved.
 
 set -e -u -x
 
+if [ "${SRCROOT}" = "" ]; then
+    echo "This script is intended to be called by Xcode during a build."
+    exit 1
+fi
+
 # srcdir is ., makes things easy
 cd "${SRCROOT}/extensions"
 
 # stick target-dir in T, keeps things short
-T="${BUILT_PRODUCTS_DIR}/${UNLOCALIZED_RESOURCES_FOLDER_PATH}/extensions/mjolnir"
+T="${BUILT_PRODUCTS_DIR}/${UNLOCALIZED_RESOURCES_FOLDER_PATH}/extensions/hammerspoon"
 
 rm -rf "${T}"
